+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Move BigEndian helper functions in `tracetranslator` to an internal package.(#3298)
 - Rename `configtest.LoadConfigFile` to `configtest.LoadConfigAndValidate` (#3306)
 - Replace `ExtensionCreateParams` with `ExtensionCreateSettings` (#3294)
+- Replace `ExporterCreateParams` with `ExporterCreateSettings` (#3164)
 
 ## 💡 Enhancements 💡